# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Psych Manager Origin API",
        "contact": {
            "email": "info@ps.ru"
        },
        "version": "1.0 beta"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api/v1",
            "description": "PMO server"
        }
    ],
    "paths": {
...
        "/login-side": {
            "post": {
                "tags": [
                    "oAuth"
                ],
                "summary": "Authentication by email",
                "description": "\r\n\r\n    ### Example URI\r\n     **POST** https://your-website.com/api/v1/login-side",
                "operationId": "oauth2.login.email",
                "parameters": [
                    {
                        "name": "X-Localization",
                        "in": "header",
                        "description": "Lang code",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "de"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    "username",
                                    "password",
                                    "grant_type",
                                    "client_id",
                                    "client_secret"
                                ],
                                "properties": {
                                    "username": {
                                        "description": "User email.",
                                        "type": "string",
                                        "example": "some_user@flywert.com"
                                    },
                                    "password": {
                                        "description": "User password.",
                                        "type": "string",
                                        "example": "some_user@flywert.com"
                                    },
                                    "grant_type": {
                                        "description": "oAuth grant type",
                                        "type": "string",
                                        "example": "password"
                                    },
                                    "client_id": {
                                        "description": "oAuth client id",
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "client_secret": {
                                        "description": "oAuth client secret",
                                        "type": "string",
                                        "example": "CyQaejvE9Tq2ykXW1aCz4aYpxU8OEpJngkVWjpHj"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "**OK** Successful auth login",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token_type": {
                                            "description": "Authorize a client to access the user's account.",
                                            "type": "string"
                                        },
                                        "expires_in": {
                                            "description": "Authorize a client to access the user's account.",
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "access_token": {
                                            "description": "Authorize a client to access the user's account.",
                                            "type": "string"
                                        },
                                        "refresh_token": {
                                            "description": "Authorize a client to access the user's account.",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "**Unauthorized** Invalid credentials."
                    },
                    "422": {
                        "description": "**Unprocessable Entity** Required fields are missing or cannot be processed."
                    },
                    "500": {
                        "description": "**Server Errors** Could not create token."
                    }
                }
            }
        },
        "/login": {
            "post": {
                "tags": [
                    "oAuth"
                ],
                "summary": "Authentication by email",
                "description": "\r\n\r\n    ### Example URI\r\n     **POST** https://your-website.com/api/v1/login",
                "operationId": "oauth2.login",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    "phone",
                                    "password",
                                    "grant_type",
                                    "client_id",
                                    "client_secret"
                                ],
                                "properties": {
                                    "email": {
                                        "description": "User email.",
                                        "type": "string",
                                        "example": "mail@qwerty213.ru39"
                                    },
                                    "password": {
                                        "description": "User password.",
                                        "type": "string",
                                        "example": "some_user@flywert.com"
                                    },
                                    "grant_type": {
                                        "description": "oAuth grant type",
                                        "type": "string",
                                        "example": "password"
                                    },
                                    "client_id": {
                                        "description": "oAuth client id",
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "client_secret": {
                                        "description": "oAuth client secret",
                                        "type": "string",
                                        "example": "CyQaejvE9Tq2ykXW1aCz4aYpxU8OEpJngkVWjpHj"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "**OK** Successful auth login",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token_type": {
                                            "description": "Authorize a client to access the user's account.",
                                            "type": "string"
                                        },
                                        "expires_in": {
                                            "description": "Authorize a client to access the user's account.",
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "access_token": {
                                            "description": "Authorize a client to access the user's account.",
                                            "type": "string"
                                        },
                                        "refresh_token": {
                                            "description": "Authorize a client to access the user's account.",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "**Unauthorized** Invalid credentials."
                    },
                    "422": {
                        "description": "**Unprocessable Entity** Required fields are missing or cannot be processed."
                    },
                    "500": {
                        "description": "**Server Errors** Could not create token."
                    }
                }
            }
        },
        "/logout-side": {
            "get": {
                "tags": [
                    "oAuth"
                ],
                "summary": "Logout and Token Invalidation",
                "description": "\r\n\r\n    Sending an request to logout endpoint with a valid API token will also invalidate that token.\r\n    ### Example URI\r\n     **GET** https://your-website.com/api/v1/logout-side",
                "operationId": "oauth2 logout",
                "responses": {
                    "200": {
                        "description": "**OK** Successful register"
                    },
                    "401": {
                        "description": "**Unauthorized** Invalid credentials."
                    }
                },
                "security": [
                    {
                        "passport": [
                            "*"
                        ]
                    }
                ]
            }
        }
    },
    "components": {
...
        "securitySchemes": {
            "X-API-KEY": {
                "type": "apiKey",
                "name": "X-API-KEY",
                "in": "header"
            }
        }
    },
    "security": [
        []
    ]
}